Split unit tests to two translation units for better categorization and shorter compile times
